pyproject.toml
[tool.poetry]
name = "itmss_api"
version = "0.1.0"
description = ""
authors = ["Renan moura <rvmoura.96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "^3.14.0"
django-filter = "^22.1"
ipython = "^8.10.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
